pyproject.toml
[project]
name = "linkedinese"
version = "0.1.0"
description = "Translate between plain English and LinkedIn Speak."
authors = [{ name = "Michael Tang" }]
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.1",
    "slowapi>=0.1.9",
]

[tool.uv]
package = false
